+++ ws2lzfrontend/lang/lang_en_US.ts
@@ -4,62 +4,80 @@
 <context>
     <name>main</name>
     <message>
-        <location filename="../src/ws2lzfrontend/main.cpp" line="43"/>
-        <source>Input path to the XML configuration file.</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <location filename="../src/ws2lzfrontend/main.cpp" line="43"/>
-        <source>configuration file</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
         <location filename="../src/ws2lzfrontend/main.cpp" line="44"/>
-        <source>Output path to an uncompressed LZ file.</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <location filename="../src/ws2lzfrontend/main.cpp" line="44"/>
-        <source>uncompressed output file</source>
+        <source>Input path to the XML configuration file.
+Cannot be used with -r</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../src/ws2lzfrontend/main.cpp" line="45"/>
-        <source>Output path to a compressed LZ file.</source>
+        <source>config file</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../src/ws2lzfrontend/main.cpp" line="45"/>
+        <location filename="../src/ws2lzfrontend/main.cpp" line="49"/>
         <source>output file</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../src/ws2lzfrontend/main.cpp" line="46"/>
-        <source>The version of SMB to generate an LZ file for (1/2/deluxe).</source>
+        <source>Input path to a stagedef (.stagedef / .lz.raw) to use as a stage configuration.
+This will automatically detect whether an SMB 1 or SMB 2 config is being used.
+Cannot be used with -c</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../src/ws2lzfrontend/main.cpp" line="46"/>
+        <location filename="../src/ws2lzfrontend/main.cpp" line="48"/>
+        <source>stagedef file</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../src/ws2lzfrontend/main.cpp" line="49"/>
+        <source>Output path to a compressed stagedef (.lz) file.</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../src/ws2lzfrontend/main.cpp" line="50"/>
+        <source>Output path to a stagedef file (.stagedef / .lz.raw).</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../src/ws2lzfrontend/main.cpp" line="50"/>
+        <source>stagedef</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../src/ws2lzfrontend/main.cpp" line="51"/>
         <source>version</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../src/ws2lzfrontend/main.cpp" line="67"/>
+        <location filename="../src/ws2lzfrontend/main.cpp" line="51"/>
+        <source>The version of SMB to generate an stagedef file for (1/2/deluxe).</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../src/ws2lzfrontend/main.cpp" line="72"/>
         <source>No confiuration file specified. Use --help for more info.</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../src/ws2lzfrontend/main.cpp" line="73"/>
+        <location filename="../src/ws2lzfrontend/main.cpp" line="75"/>
+        <source>Cannot specify both an XML config and a raw LZ config. Use --help for more info.</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../src/ws2lzfrontend/main.cpp" line="81"/>
         <source>No output file specified. Use --help for more info.</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../src/ws2lzfrontend/main.cpp" line="79"/>
+        <location filename="../src/ws2lzfrontend/main.cpp" line="87"/>
         <source>No game version specified - defaulting to Super Monkey Ball 2 - Use --game-version to specify a different one</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../src/ws2lzfrontend/main.cpp" line="91"/>
+        <location filename="../src/ws2lzfrontend/main.cpp" line="99"/>
         <source>Invalid game version specified. Use --help for more info.</source>
         <translation type="unfinished"></translation>
     </message>